requirements.txt
scapy
pycryptodome
pycryptodomex